requirements.txt
python-dotenv
pandas 
numpy 
pymongo 
certifi
pymongo[srv]
scikit-learn
dill 
pyaml
mlflow
dagshub
fastapi
uvicorn
# -e .